@@ -7,29 +7,27 @@
         "watch-poll": "npm run watch -- --watch-poll",
         "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
         "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --hide-modules --progress --config=node_modules/laravel-mix/setup/webpack.config.js"
     },
     "devDependencies": {
-        "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+        "@babel/plugin-syntax-dynamic-import": "^7.8.0",
+        "@types/leaflet": "^1.5.7",
         "axios": "^0.18",
-        "bootstrap": "^4.1.0",
+        "bootstrap": "^4.4.1",
         "cross-env": "^5.2.1",
-        "jquery": "^3.2",
-        "laravel-mix": "^4.1.2",
-        "lodash": "^4.17.15",
-        "popper.js": "^1.12",
-        "resolve-url-loader": "^2.3.1",
-        "sass": "^1.22.10",
-        "sass-loader": "^7.3.1",
-        "vue": "^2.5.17"
-    },
-    "dependencies": {
-        "@types/leaflet": "^1.5.1",
-        "leaflet": "^1.5.1",
+        "jquery": "^3.4.1",
+        "laravel-mix": "^4.1.4",
+        "leaflet": "^1.6.0",
+        "popper.js": "^1.16.0",
+        "resolve-url-loader": "^3.1.1",
+        "sass": "^1.24.4",
+        "sass-loader": "7.1.0",
+        "vue": "^2.6.11",
         "vue-slicksort": "^1.1.3",
-        "vue-template-compiler": "^2.6.10",
+        "vue-template-compiler": "^2.6.11",
         "vue-youtube-embed": "^2.2.2",
-        "vue2-leaflet": "^2.2.1",
-        "vuex": "^3.1.1"
-    }
+        "vue2-leaflet": "^2.4.2",
+        "vuex": "^3.1.2"
+    },
+    "dependencies": {}
 }
