@@ -20,9 +20,7 @@
                     <input v-else
                            placeholder="Insert media link"
                            class="formRightInputUrl">
-                    <form action="/setDataMap/uploadImg" method="post" enctype="multipart/form-data" class="formRightUploadFile">
-                        <slot>
-                        </slot>
+                    <form action="/uploadImg" method="post" enctype="multipart/form-data" class="formRightUploadFile">
                         <label class="formRightUploadButton">
                             <img src="/images/cloud-upload.png">Upload from drive
                             <input type="file" multiple name="image" onchange="this.form.submit()"
@@ -34,14 +32,14 @@
                     <template v-if="checkExistImages">
                         <img :src="selectedEvent.mediaUrl"
                              alt="image"
-                             width="270"
-                             height="144"
+                             width="340"
+                             height="176"
                              class="emptyMediaBlock">
                     </template>
                     <template v-else-if="selectedEvent.mediaUrl.indexOf('.youtu') !== -1">
                         <youtube :video-id="getYouTubeIdOfSelectedEvent"
-                                 :player-width="270"
-                                 :player-height="144"
+                                 :player-width="260"
+                                 :player-height="140"
                                  class="emptyMediaBlock">
                         </youtube>
                     </template>
@@ -132,106 +130,93 @@
     }
 </script>
 
-<style scoped>
-    .form {
-        display: flex;
-        margin-top: 10px;
-    }
+<style lang="sass" scoped>
+    .form
+        display: flex
+        margin-top: 10px
 
-    .formLeft {
-        display: flex;
-        flex-direction: column;
-        margin-right: 1%;
-        width: 46%;
-    }
+    .formLeft
+        display: flex
+        flex-direction: column
+        margin-right: 1%
+        width: 46%
 
-    .inputHeader {
-        font-size: 24px;
-        margin-bottom: 2%;
-        padding: 3px;
-        height: 28px;
-        border: 1px solid #4d565661;
-        border-radius: 4px;
-    }
+    .inputHeader
+        font-size: 24px
+        margin-bottom: 2%
+        padding: 3px
+        height: 28px
+        border: 1px solid #4d565661
+        border-radius: 4px
 
-    .inputTextarea {
-        font-size: 18px;
-        padding: 3px;
-        flex-grow: 1;
-        resize: none;
-        border: 1px solid #4d565661;
-        border-radius: 4px;
-        font-family: 'Turnip RE', Georgia, 'Times New Roman', Times, serif;
-    }
+    .inputTextarea
+        font-size: 18px
+        padding: 3px
+        flex-grow: 1
+        resize: none
+        border: 1px solid #4d565661
+        border-radius: 4px
+        font-family: 'Turnip RE', Georgia, 'Times New Roman', Times, serif
 
-    .formRight {
-        display: flex;
-        flex-wrap: wrap;
-        width: 53%;
-        min-height: 166px;
-        border: 1px solid #4d565661;
-        border-radius: 4px;
-        background: #f8f8f8;
-    }
+    .formRight
+        display: flex
+        flex-wrap: wrap
+        width: 53%
+        min-height: 180px
+        border: 1px solid #4d565661
+        border-radius: 4px
+        background: #f8f8f8
 
-    .formRightContent {
-        display: flex;
-        flex-wrap: wrap;
-        margin: auto;
-        max-width: 540px;
-        width: 100%;
-    }
+    .formRightContent
+        display: flex
+        flex-wrap: wrap
+        margin: auto
+        max-width: 540px
+        width: 100%
 
-    .formRightInputUrl {
-        font-size: 20px;
-        margin: 4% 2% 2% 2%;
-        width: 100%;
-    }
+    .formRightInputUrl
+        font-size: 20px
+        margin: 4% 2% 2% 2%
+        width: 100%
 
-    .formRightUploadFile {
-        display: flex;
-        width: 100%;
-    }
+    .formRightUploadFile
+        display: flex
+        width: 100%
 
-    .emptyMediaBlock {
-        width: 270px;
-        height: 144px;
-        display: flex;
-        padding: 8px;
-        margin: auto;
-    }
+    .emptyMediaBlock
+        width: 270px
+        height: 144px
+        display: flex
+        padding: 8px
+        margin: auto
 
-    .emptyMediaBlock p {
-        font-size: 26px;
-        text-align: center;
-        background-color: #dddddd;
-        color: #777777;
-        line-height: 144px;
-        width: 100%;
-        margin: auto;
-        font-family: "Helvetica Neue", Helvetica, sans-serif;
-    }
+        p
+            font-size: 26px
+            text-align: center
+            background-color: #dddddd
+            color: #777777
+            line-height: 144px
+            width: 100%
+            margin: auto
+            font-family: "Helvetica Neue", Helvetica, sans-serif
 
-    .formRightUploadButton {
-        font-size: 18px;
-        margin: 2%;
-        padding: 6px 0;
-        background: none;
-        width: 100%;
-        border: 1px dashed #4d565661;
-        transition: all 0.2s;
-        cursor: pointer;
-    }
+    .formRightUploadButton
+        font-size: 18px
+        margin: 2%
+        padding: 6px 0
+        background: none
+        width: 100%
+        border: 1px dashed #4d565661
+        transition: all 0.2s
+        cursor: pointer
 
-    .formRightUploadButton:hover {
-        background-color: #FFFFFF;
-        border: 1px solid #4d565661;
-    }
+        &:hover
+            background-color: #FFFFFF
+            border: 1px solid #4d565661
 
-    .formRightUploadButton img {
-        height: 18px;
-        position: relative;
-        bottom: -3px;
-        margin-right: 4px;
-    }
+        img
+            height: 18px
+            position: relative
+            bottom: -3px
+            margin-right: 4px
 </style>
