@@ -1,5 +1,7 @@
 <template>
-
+    <div>
+        <h1>Coming soon</h1>
+    </div>
 </template>
 
 <script>
